fix(gap-sweep): address Copilot bot inline review (PR #511)
Three inline thread fixes:

1. **Clear read-config indicators on early-return paths.** LoadListFromPointer
   previously returned early on `rom == null`, `ptr == 0`, or non-safety
   offset without resetting BeforeBaseAddr/AfterBaseAddr and the matching
   read counts. Stale values from a prior successful load could leak into
   the top read-config bar after a failed reload. Add a ResetReadConfig
   helper called on all three early-return paths.

2. **GetRawRomReport BeforePtr/AfterPtr now use p32, not u32.** The VM's
   BeforeEventPointer / AfterEventPointer properties are loaded via p32
   (offset form, mask stripped) per the WF P0 convention. GetRawRomReport
   was reporting the raw u32 (pointer form, mask on), which made the
   DataVerifiableSweep FieldLevelCrossCheck see ~0x08000000 drift between
   the two reports for the per-row pointers. Switch to p32 for symmetry.

3. **Cast CopyTo dest offsets to int explicitly.** Although uint -> long
   implicit conversion makes the existing CopyTo(byte[], long) overload
   bind, the Copilot reviewer flagged the uint indexes as ambiguous.
   Adding explicit `(int)` casts removes the ambiguity and makes intent
   clearer.

Parity tests all 15/15 pass; DataVerifiableSweep 1193/1193 pass.
Per-row round-trip tests (already in place) continue to pin the
p32/write_p32 contract.

+++ FEBuilderGBA.Avalonia.Tests/GapSweep/WorldMapEventPointerParityTests.cs
@@ -551,24 +551,24 @@ public class WorldMapEventPointerParityTests
         {
             for (int i = 0; i < stageClearSlots.Length; i++)
             {
-                BitConverter.GetBytes(stageClearSlots[i]).CopyTo(bytes, baseClear + (uint)(i * 4));
+                BitConverter.GetBytes(stageClearSlots[i]).CopyTo(bytes, (int)(baseClear + (uint)(i * 4)));
             }
             // Plant a sentinel NULL after the slots so the loader has
             // somewhere to stop even if the test array doesn't terminate.
-            BitConverter.GetBytes(0u).CopyTo(bytes, baseClear + (uint)(stageClearSlots.Length * 4));
+            BitConverter.GetBytes(0u).CopyTo(bytes, (int)(baseClear + (uint)(stageClearSlots.Length * 4)));
             BitConverter.GetBytes(baseClear | 0x08000000u).CopyTo(bytes,
-                rom.RomInfo.worldmap_event_on_stageclear_pointer);
+                (int)rom.RomInfo.worldmap_event_on_stageclear_pointer);
         }
 
         if (stageSelectSlots != null)
         {
             for (int i = 0; i < stageSelectSlots.Length; i++)
             {
-                BitConverter.GetBytes(stageSelectSlots[i]).CopyTo(bytes, baseSelect + (uint)(i * 4));
+                BitConverter.GetBytes(stageSelectSlots[i]).CopyTo(bytes, (int)(baseSelect + (uint)(i * 4)));
             }
-            BitConverter.GetBytes(0u).CopyTo(bytes, baseSelect + (uint)(stageSelectSlots.Length * 4));
+            BitConverter.GetBytes(0u).CopyTo(bytes, (int)(baseSelect + (uint)(stageSelectSlots.Length * 4)));
             BitConverter.GetBytes(baseSelect | 0x08000000u).CopyTo(bytes,
-                rom.RomInfo.worldmap_event_on_stageselect_pointer);
+                (int)rom.RomInfo.worldmap_event_on_stageselect_pointer);
         }
 
         // Re-load so RomInfo (already populated by the first LoadLow) sees
